embed CombinationController action into product view to render prototype
